Implement grid extraction
We've chosen to represent the grid as a map with markers for each
quadrant of the grid. This allows us to fetch subgroups in an easier way
in the future, since that is the most complicated part to build a logic
for.

With this representation, the complexity gets pushed over to the columns.